requirements.txt
-f /usr/share/pip-wheels
Babel==2.1.1
Django==1.8.7
django-phonenumber-field==1.0.0
phonenumbers==7.1.1
pytz==2015.7
wheel==0.24.0
